+++ lib/src/perf_test_proc.dart
@@ -5,7 +5,19 @@ import 'package:perf_test/perf_test.dart';
 
 /// Actual user-defined procedure to test
 ///
-typedef PerfTestOneProc = void Function(PerfTestOne test, int lapNo);
+typedef PerfTestOneProc = Future Function(PerfTestOne test, int lapNo);
+
+/// Convenience definition, similar to PerfTestOneProc
+///
+typedef PerfTestProc = Future Function(PerfTestOne test, int lapNo);
+
+/// Actual user-defined procedure to test
+///
+typedef PerfTestOneProcSync = void Function(PerfTestOne test, int lapNo);
+
+/// Convenience definition, similar to PerfTestOneProcSync
+///
+typedef PerfTestProcSync = void Function(PerfTestOne test, int lapNo);
 
 /// User-defined procedure to show lot result
 ///